Setup ecto and add analytics module to store the secrets count (#33)
The PR ended up a bit large because I didn't want to maintain multiple branches and merge this in batches. As soon as the review app works, we merge it to keep the migration easy.

This PR sets up Ecto and adds all associated boilerplate for the Ecto install for a Phoenix project.

It also implements a KV store and Analytics module to store various types of analytic values. It also integrates with the rest of the flow by automatically incrementing the secrets counter value when effectively adding new secrets.

**Currently there is no way to query for the analytics value without visiting the console, we can add that later on after discussing with @rafaelferreira6 where we put this info.**